accessibility, added custom gradient component modified from ui belick

+++ app/_lib/helpers.ts
@@ -22,10 +22,10 @@ export function formatDate(date: string, includeRelative = false) {
 	}
 
 	const fullDate = targetDate.toLocaleString("en-us", {
-		month: "long",
-		day: "numeric",
-		year: "numeric",
-	});
+  month: "numeric",
+  day: "numeric",
+  year: "numeric",
+});
 
 	if (!includeRelative) {
 		return fullDate;
